Выполняет разбивку на шаблоны
