+++ templates/index.php
@@ -1,0 +1,31 @@
+<h2 class="content__main-heading">Список задач</h2>
+
+<form class="search-form" action="index.php" method="post">
+    <input class="search-form__input" type="text" name="" value="" placeholder="Поиск по задачам">
+
+    <input class="search-form__submit" type="submit" name="" value="Искать">
+</form>
+
+<div class="tasks-controls">
+    <nav class="tasks-switch">
+        <a href="/" class="tasks-switch__item tasks-switch__item--active">Все задачи</a>
+        <a href="/" class="tasks-switch__item">Повестка дня</a>
+        <a href="/" class="tasks-switch__item">Завтра</a>
+        <a href="/" class="tasks-switch__item">Просроченные</a>
+    </nav>
+
+    <label class="checkbox">
+        <input
+            class="checkbox__input visually-hidden show_completed"
+            type="checkbox"
+            <?php if ($show_complete_tasks === 1):?>checked<?php endif; ?>
+        />
+        <span class="checkbox__text">Показывать выполненные</span>
+    </label>
+</div>
+
+<table class="tasks">
+    <?php foreach ($visible_tasks as $task): ?>
+        <?= include_template(VIEWS_PATH . 'partials/task.php', ['task' => $task]); ?>
+    <?php endforeach ?>
+</table>
